Package RhinoToSlicer as a Rhino Python plug-in
