+++ src/RhinoToSlicer/__init__.py
@@ -1,0 +1,4 @@
+"""RhinoToSlicer Rhino Python plug-in."""
+
+__all__ = ["__version__"]
+__version__ = "0.2.0"
