+++ src/RhinoToSlicer/commands/__init__.py
@@ -1,0 +1,5 @@
+"""Command implementations for the RhinoToSlicer plug-in."""
+
+from .send_to_prusa import RunCommand, __commandname__  # noqa: F401
+
+__all__ = ["RunCommand", "__commandname__"]
